+++ platform/linux/templates/sdcc-mcs51/main.c
@@ -1,0 +1,7 @@
+#include <8051.h>
+
+void main(void) {
+	for(;;) {
+		P1--;
+	}
+}